\name{print.EBLUP}
\alias{print.EBLUP}
\title{
print the value of EBLUP predictor}

\usage{\method{print}{EBLUP}(x, ...)
}

\arguments{
\item{x}{the object of class 'EBLUP'.}
\item{\dots}{not used.}
}

\description{Print the value of EBLUP predictor.}


\author{
Alicja Wolny-Dominiak}
