\name{print.ebpLMMne}
\alias{print.ebpLMMne}

\title{
print the value of ebpLMMne predictor}

\usage{\method{print}{ebpLMMne}(x, ...)}

\arguments{
\item{x}{the object of class 'ebpLMMne'}
\item{\dots}{not used}
}

\description{Print the value of ebpLMMne predictor.}

\author{
Alicja Wolny-Dominiak}
