\name{print.plugInLMM}
\alias{print.plugInLMM}

\title{
print the value of plugInLMM predictor}

\usage{\method{print}{plugInLMM}(x, ...)
}

\description{Print the value of plugInLMM predictor.}

\arguments{
\item{x}{the object of class 'plugInLMM'}
\item{\dots}{not used}
}

\author{
Alicja Wolny-Dominiak}

