\name{summary.EBLUP}
\alias{summary.EBLUP}
\title{
Summary of EBLUP prediction}

\usage{\method{summary}{EBLUP}(object, ...)}

\arguments{
\item{object}{the object of class 'EBLUP'.}
\item{\dots}{not used.}
}

\description{Print the summary of EBLUP prediction and LMM model.}

\author{
Alicja Wolny-Dominiak}
