\name{summary.ebpLMMne}
\alias{summary.ebpLMMne}

\title{
Summary of ebpLMMne prediction}

\usage{\method{summary}{ebpLMMne}(object, ...)}

\arguments{
\item{object}{the object of class 'ebpLMMne'}
\item{\dots}{not used}
}

\description{Print the summary of ebpLMMne prediction and LMM.}


\author{
Alicja Wolny-Dominiak}
