\name{summary.plugInLMM}
\alias{summary.plugInLMM}

\title{
Summary of plugInLMM prediction
}

\usage{\method{summary}{plugInLMM}(object, ...)}

\arguments{
\item{object}{the object of class 'plugInLMM'.}
\item{\dots}{not used.}
}

\description{Print the summary of ebpLMMne prediction and LMM model.}


\author{
Alicja Wolny-Dominiak}

